Fix for issue #12 - accented characters and emoji causing crash
In several places we retrieve nested data from the database as
utf-8 json, which is then parsed into a native perl data structure.
But the database driver itself handles the decoding from utf-8
into internal perl encoding. If we try to repeat that decoding
when calling from_json() there's a crash. So we now tell
from_json() not to do that!